+++ console/src/main/java/org/nervos/ckb/utils/SignUtils.java
@@ -1,7 +1,8 @@
 package org.nervos.ckb.utils;
 
-import org.bouncycastle.jcajce.provider.digest.SHA3;
+import org.nervos.ckb.crypto.Blake2b;
 import org.nervos.ckb.crypto.ECKeyPair;
+import org.nervos.ckb.crypto.Hash;
 import org.nervos.ckb.crypto.Sign;
 import org.nervos.ckb.methods.type.CellInput;
 import org.nervos.ckb.methods.type.CellOutput;
@@ -17,22 +18,22 @@ public class SignUtils {
 
     public static List<CellInput> signSigHashAllInputs(List<CellInput> cellInputs, List<CellOutput> cellOutputs, String privateKey) {
         String sigHashType = "1";
-        SHA3.Digest256 sha3 = new SHA3.Digest256();
-        sha3.update(sigHashType.getBytes());
+        Blake2b blake2b = new Blake2b();
+        blake2b.update(sigHashType.getBytes());
         for (CellInput cellInput : cellInputs) {
-            sha3.update(Numeric.hexStringToByteArray(cellInput.previousOutput.hash));
-            sha3.update((String.valueOf(cellInput.previousOutput.index)).getBytes());
-            sha3.update(Numeric.hexStringToByteArray(cellInput.unlock.getTypeHash()));
+            blake2b.update(Numeric.hexStringToByteArray(cellInput.previousOutput.hash));
+            blake2b.update((String.valueOf(cellInput.previousOutput.index)).getBytes());
+            blake2b.update(Numeric.hexStringToByteArray(cellInput.unlock.getTypeHash()));
         }
 
         for (CellOutput cellOutput : cellOutputs) {
-            sha3.update((String.valueOf(cellOutput.capacity)).getBytes());
-            sha3.update(Numeric.hexStringToByteArray(cellOutput.lock));
+            blake2b.update((String.valueOf(cellOutput.capacity)).getBytes());
+            blake2b.update(Numeric.hexStringToByteArray(cellOutput.lock));
             if (cellOutput.type != null) {
-                sha3.update(Numeric.hexStringToByteArray(cellOutput.type.getTypeHash()));
+                blake2b.update(Numeric.hexStringToByteArray(cellOutput.type.getTypeHash()));
             }
         }
-        byte[] messageHash = sha3.digest();
+        byte[] messageHash = blake2b.doFinalBytes();
 
         for (CellInput cellInput : cellInputs) {
             List<String> args = new ArrayList<>();