+++ utils/src/main/java/org/nervos/ckb/utils/FileUtils.java
@@ -1,8 +1,10 @@
 package org.nervos.ckb.utils;
 
 import java.io.BufferedReader;
+import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
+import java.nio.file.Files;
 
 /**
  * Created by duanyytop on 2019-01-29.
@@ -32,6 +34,16 @@ public class FileUtils {
             }
         }
         return null;
+    }
+
+    public static byte[] readFileForBytes(String path) {
+        File file = new File(path);
+        try {
+            return Files.readAllBytes(file.toPath());
+        } catch (IOException e) {
+            e.printStackTrace();
+            return null;
+        }
     }